[AAP-78775] Move diagnose_key tests into test_jwt_token.py
